Changes to ball actuator
Deleted all references to the arm because we are no longer using it. Added the correct button values that the drivers wanted. Made it so both of the joystick can control the movement and turning speed. The Actuator, Ball Servo, and Drive train in this branch are ready for competirion.